scores, student detail

+++ src/student/student.controller.ts
@@ -19,18 +19,21 @@ import {
   INSTRUCTOR_DOES_NOT_EXIST,
   SOMETHING_WENT_WRONG,
   STUDENT_ALREADY_EXIST,
+  UNAUTHORIZED_INSTRUCTOR_REQUEST,
 } from "src/utilities/errors";
 import { UpdateStudentDto } from "./dto/update-student.dto";
 import { IdParam } from "src/utilities/decorators/paramId.decorator";
 import { InstructorService } from "src/instructor/instructor.service";
 import { QuizService } from "src/quiz/quiz.service";
+import { ScoreService } from "src/score/score.service";
 
 @Controller("students")
 export class StudentController {
   constructor(
     private readonly studentService: StudentService,
     private readonly instructorService: InstructorService,
-    private readonly quizService: QuizService
+    private readonly quizService: QuizService,
+    private readonly scoreService: ScoreService
   ) {}
 
   @Get()
@@ -79,6 +82,16 @@ export class StudentController {
     @User("_id") instructorId: string,
     @Body() studentDto: UpdateStudentDto
   ) {
+    const isInstructorAuthorized =
+      await this.studentService.checkInstructorAuthForStudent(
+        instructorId,
+        studentDto._id
+      );
+
+    if (!isInstructorAuthorized) {
+      throw new ExceptionBadRequest(UNAUTHORIZED_INSTRUCTOR_REQUEST);
+    }
+
     const foundStudent = await this.studentService.findOneByNickName(
       studentDto.nickname,
       studentDto._id
@@ -100,6 +113,16 @@ export class StudentController {
     @IdParam() studentId: string
   ) {
     try {
+      const isInstructorAuthorized =
+        await this.studentService.checkInstructorAuthForStudent(
+          instructorId,
+          studentId
+        );
+
+      if (!isInstructorAuthorized) {
+        throw new ExceptionBadRequest(UNAUTHORIZED_INSTRUCTOR_REQUEST);
+      }
+
       await this.studentService.delete(studentId);
 
       const includedQuizzes = await this.quizService.findByStudentId(studentId);
@@ -122,5 +145,30 @@ export class StudentController {
   @UseGuards(UserGuard)
   async profile(@User("_id") studentId: string) {
     return await this.studentService.getProfile(studentId);
+  }
+
+  @Get(":_id")
+  @UseGuards(UserGuard)
+  async detail(
+    @User("_id") instructorId: string,
+    @IdParam() studentId: string,
+    @Query() query: PaginationQueryDto
+  ) {
+    const isInstructorAuthorized =
+      await this.studentService.checkInstructorAuthForStudent(
+        instructorId,
+        studentId
+      );
+
+    if (!isInstructorAuthorized) {
+      throw new ExceptionBadRequest(UNAUTHORIZED_INSTRUCTOR_REQUEST);
+    }
+
+    const student = await this.studentService.getProfile(studentId);
+    const scores = await this.scoreService.paginate({
+      ...query,
+      student: studentId,
+    });
+    return { ...student, scores };
   }
 }
